requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.15.2
pillow==12.0.0
sqlparse==0.5.3
